서버 엔트리 구조 분리 및 환경변수 정리
