+++ .env.example
@@ -1,0 +1,8 @@
+PORT=3000
+SERVER_BASE_URL=http://localhost:3000
+
+DB_HOST=localhost
+DB_USER=root
+DB_PASSWORD=your_password
+DB_NAME=your_database
+DB_PORT=3306
