+++ server/package.json
@@ -14,7 +14,9 @@
   "dependencies": {
     "body-parser": "^2.2.0",
     "cors": "^2.8.5",
+    "dotenv": "^17.4.2",
     "express": "^5.1.0",
+    "multer": "^2.2.0",
     "mysql2": "^3.14.1"
   }
 }
